+++ integration_tests/sdk/data_integration_tests/conftest.py
@@ -67,6 +67,6 @@ def flow_manager(client, flow_name, engine):
 
 
 def pytest_sessionfinish(session, exitstatus):
-    if not hasattr(session.config, "workerinput"):
+    if not hasattr(session.config, "workerinput") and not session.config.getoption("keep_flows"):
         client = aq.Client(*get_aqueduct_config())
         delete_all_flows(client)
